[oc/swift cross resolve] split ObjcExportHeaderGenerator into several files
