+++ backend.native/compiler/ir/backend.native/src/org/jetbrains/kotlin/backend/konan/objcexport/objcTypes.kt
@@ -1,0 +1,98 @@
+/*
+ * Copyright 2010-2017 JetBrains s.r.o.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jetbrains.kotlin.backend.konan.objcexport
+
+internal sealed class ObjCType {
+    final override fun toString(): String = this.render()
+
+    abstract fun render(attrsAndName: String): String
+
+    fun render() = render("")
+
+    protected fun String.withAttrsAndName(attrsAndName: String) =
+            if (attrsAndName.isEmpty()) this else "$this ${attrsAndName.trimStart()}"
+}
+
+internal sealed class ObjCReferenceType : ObjCType()
+
+internal sealed class ObjCNonNullReferenceType : ObjCReferenceType()
+
+internal data class ObjCNullableReferenceType(val nonNullType: ObjCNonNullReferenceType) : ObjCReferenceType() {
+    override fun render(attrsAndName: String) = nonNullType.render(" _Nullable".withAttrsAndName(attrsAndName))
+}
+
+internal class ObjCClassType(
+        val className: String,
+        val typeArguments: List<ObjCNonNullReferenceType> = emptyList()
+) : ObjCNonNullReferenceType() {
+
+    override fun render(attrsAndName: String) = buildString {
+        append(className)
+        if (typeArguments.isNotEmpty()) {
+            append("<")
+            typeArguments.joinTo(this) { it.render() }
+            append(">")
+        }
+        append(" *")
+        append(attrsAndName)
+    }
+}
+
+internal class ObjCProtocolType(val protocolName: String) : ObjCNonNullReferenceType() {
+
+    override fun render(attrsAndName: String) = "id<$protocolName>".withAttrsAndName(attrsAndName)
+}
+
+internal object ObjCIdType : ObjCNonNullReferenceType() {
+    override fun render(attrsAndName: String) = "id".withAttrsAndName(attrsAndName)
+}
+
+internal object ObjCInstanceType : ObjCNonNullReferenceType() {
+    override fun render(attrsAndName: String): String = "instancetype".withAttrsAndName(attrsAndName)
+}
+
+internal class ObjCBlockPointerType(
+        val returnType: ObjCReferenceType, val parameterTypes: List<ObjCReferenceType>
+) : ObjCNonNullReferenceType() {
+
+    override fun render(attrsAndName: String) = returnType.render(buildString {
+        append("(^")
+        append(attrsAndName)
+        append(")(")
+        if (parameterTypes.isEmpty()) append("void")
+        parameterTypes.joinTo(this) { it.render() }
+        append(')')
+    })
+}
+
+internal class ObjCPrimitiveType(val cName: String) : ObjCType() {
+    override fun render(attrsAndName: String) = cName.withAttrsAndName(attrsAndName)
+}
+
+internal class ObjCPointerType(val pointee: ObjCType, val nullable: Boolean = false) : ObjCType() {
+    override fun render(attrsAndName: String) =
+            pointee.render("*${if (nullable) {
+                " _Nullable".withAttrsAndName(attrsAndName)
+            } else {
+                attrsAndName
+            }}")
+}
+
+internal object ObjCVoidType : ObjCType() {
+    override fun render(attrsAndName: String) = "void".withAttrsAndName(attrsAndName)
+}
+
